SlotRoleRegistry: role names are case insensitive
This ensures that slot role names are consistently treated
as case insensitive. Previously, role names were converted to
lower case when stored in the database, but not when registering
or looking up a handler. This could result in spurious errors,
e.g. when the name used for registration was not lower case.

Change-Id: I15db141266d7f4efce3da54dd661b3418a7bb41d