upload: Fix incomplete type hint in UploadBase::initializePathInfo
The $fileSize parameter can be null.

I'm also removing all sudo documentation that doesn't add any
information beyond what the method and parameter names already literally
say. If code is self-documenting, repeating that does not make it easier
to read, it makes it *harder*. There is more text to read, but not more
information. The text needs to be updated, and will be wrong if forgotten.

Change-Id: I6d3fcd3aba098826d119454df923e3b1c4cc83aa