languages: Don't assume $wgUser in LanguageConverter
LanguageConverter::getPreferredVariant() may be called during session
initialization, triggered by a call to User::isUsableName(). If that
happens, the $wgUser variable cannot be used since it hasn't been
initialized at this point.

The code previously called $wgUser->isSafeToLoad() to make sure the
User object can be initialized. It however made the assumption that
the User object itself is already present, which is not necessarily the
case.

Bug: T235360
Change-Id: I62bd09d7d09b4d82d251fc6010f0dcc11f21c047