JsonContent: align object key cells at the top
The default middle vertical align can make it hard to find the object
key for long tables (where the value is itself a long object or array).
Move it to the top, and add position: sticky to make it stay on screen
as the top of the table is scrolled out of view. (position: sticky is
not supported in Internet Explorer, but in that case this should
gracefully degrade to keeping the object key at the top, which is no
worse than the previous behavior, where the object key might be out of
view in all browsers.)

The extra <span> is necessary because putting position: sticky on the
table cell itself doesn’t work.

Change-Id: Ib38710e5cc2b75971a26eabf7f144378790d3661