TemplateParser: Change cache key and version cache
Needed because we upgraded to a newer version of lightncandy, which
contains breaking changes. Cached compilations from the old versions
won't work with the new version.

Add a cache version variable so that we can bump this more easily in the
future when upgrading lightncandy.

Follows-up 7c4e9f790a48bedc1c314162a5fd2068faae1cb1

Change-Id: I7c05019df7eb044cff7bb8e67b2835970e1f384d