MultiHttpClient: Reset response headers on redirect
We set Curl to follow the redirect with CURLOPT_FOLLOWLOCATION,
however the response headers are being concatenated into a
comma-separated list. Curl follows redirect silently under the hook,
so we end up with a combination of headers from a redirect response
and an after-redirect response. We need to reset the headers.

Change-Id: I9a0f208e5479e5c126d36d788c545a23ff421438