mw.special.preferences: Allow filling in real time zone instead of offset
The "Time zone / Fill in from browser" option actually only filled in
the time offset, rather than the named time zone.

The time offset is just the number of minutes of difference from UTC,
while the time zone includes information about two (or more) time
offsets, which are used when switching to / from daylight saving time.
Using time offset instead is incorrect in large parts of the world
that use DST.

When this feature was originally written (in 2003: 84dc7809 / r1328),
JavaScript had no API to get the named time zone. But now it has!

Bug: T18763
Bug: T132264
Change-Id: I461f4f283a098effcf17cf151a9d5f9ac80c663d