UsersMultiselectWidget: Add pre-selected items as options
Any items passed via the 'selected' config can only be added as tags
if they are present in the menu, so add them as menu options too. The TitlesMultiselectWidget already does this.

Also, adapt the TitlesMultiselectWidget so that it adds to any options
passed in via the 'options' config, rather than overriding them. (The
widget already handles duplicates when adding the options to the
menu.)

This handling will be upstreamed to the MenuTagMultiselectWidget in
OOUI soon (and removed from these widgets), but we need this now to
prevent a bug, exposed by b5b9e13d55.

Bug: T243643
Change-Id: Ia52796786492d82fcd492839d3e0252f14b7f0f1