When clearing don't load the watchlist if we must clear through a job
It looks like this bug has existed in the special page since before the
time of the refactoring into WatchedItemStore although apparently
it has only just surfaced now!

This adds a new method into WatchedItemStore that decides if the
watchlist can be cleared interactively or must use the job queue.
This can then be used in the special page instead of the old logic
which would load the watchlist and then count the loaded items
(inefficient if you know your clearing the list anyway)

Bug: T243449
Change-Id: I810d89e3e1142a223430f7fc5f8598a493637a72