maintenance: Add file extension as expected for PHP scripts
The example in the block comment above uses the .php file extension
in the example but the usage string omits it and it can be misleading.

If the extension (.php) is not given, `runScript.php` will think that
the script doesn't exist and trigger a different code path rather than
actually run the script in question.

Change-Id: I4d60f65629472f2243ead2c8768f597ae23b6beb