Click on "Previous page" thumbnail shouldn't lead from page 2 to page 0
When viewing a multi-page document, "previous page" thumbnail has page=
parameter, but when previous page is page #1, this parameter is omitted
(the link points to /wiki/File:Something.pdf,
not to /w/index.php?title=File:Something.pdf&page=1). This is correct.
Not having the page parameter means page 1 will be shown.

The line of JavaScript code touched in this patch created page=0 when the
parameter was missing. Such a page does not exist. It needs to start with
1 or be omitted. Asking for page 0 can cause misbehavior depending on the
wiki's configuration.

Bug: T241747
Change-Id: I701742f539822fbe26e69201241c9c4cb291953f