Exempt users with 'ipblock-exempt' right from all cookie blocks
A cookie block is a block in the database, pointed to by a cookie.
Blocked users who switch accounts or IP addresses may continue to
be blocked due to a cookie pointing to a block against their
original account or IP address.

This extends the 'ipblock-exempt' right to exempt users with the
right from all cookie blocks. The rationale for this is:
* A cookie block against an IP does not affect any logged in users
  (as codified in BlockManager::shouldApplyCookieBlock).
* A cookie block against a user's own account does not need to be
  enforced since they will be blocked directly.
* It was decided in T233441 that a cookie block against a different
  user's account should not affect a user with the 'ipblock-exempt'
  right, to avoid users with this right from being blocked as
  collateral damage.

Bug: T233441
Change-Id: I3912b13240e0b42ad83a9527eeba844e8601d764