objectcache: improve WANObjectCache I/O spike avoidance
Rewrite checkAndSetCooloff() to have a clearer intent and a
broader scope. It now applies any time a key is evicted and
seems "risky" (based on value size and regeneration time).
Previously, only keys that used "lockTSE" were considered
for such throttling.

For scalar values, estimating the size is easy enough, though
the final object size might be smaller due to compression.
A pessimistic size is assumed for other variable types.

Bug: T239983
Change-Id: Ib233a650112a9d574a6aa744f6af23bc16d4f5d7