mediawiki.jqueryMsg: Document and test param cloning behaviour
Follows bb48ecbbc7473a.

This adds some documentation and renames the property that checks if a
non-string replacement has been used multiple times. The goal is to be
more mindful of other code that might inspect it.

A test is added to assert that the replacements are only cloned from the
second use onward, in case someone holds on to the inject object to
listen for clicks or something similar.

Change-Id: Icd1e06ed50f12483cf0884027e83dacf17eebb02