resourceloader: Optimization - avoid repeated `$this->getModule()` calls
`$this->getModule()` returns a RL object or a null. The check above already
makes sure that $module is not null so we can just set file dependencies
from the RL object returned.

In addition, this is an optimization as we reduce repeated calls to
`$this->getModule()`. Similar logic already exist in the same file from
L#116 - L#118.

Change-Id: If50c28ff5cd8eb435c9fa4f277e2d1038b52ca74