docs: Improve Doxygen aliases and map more tool annotations to @noop
For some reason, @access is outputting a literal "\n" instead of the
newline that's supposed to generate. It seems to only ever be followed
by "public", "private", or "protected", so just remove the newline.

On the other hand, "@private" and so on are sometimes followed by text.
Instead of using @access, use a custom @par to produce nicer output.

"@internal" is being used now something like "@private", to indicate
things that are public for PHP access rules but should be treated as
private. Unfortunately Doxygen has its own "@internal", so we need to
override it.

We've also started using "@unstable", which Doxygen doesn't know about.

Doxygen 1.8.17 introduces "@noop", which allows us to alias annotations
such that they do not result in strange of partial output in docs.
Wikimedia CI is now on 1.8.19.

Fix "@suppress", "@codeCoverageIgnore", and
"@codingStandardsIgnore(Start|End)" by using this.

Bug: T233788
Bug: T241182
Change-Id: Idc7ef2afa6cdf8c15f0ca5e7e96e67c3a4374a8a