widgets: Split out SearchWidgets into their own tree
Expecting people to magically know that MediaWiki\Widget\SearchInputWidget is
generic and for use on random pages, but MediaWiki\Widget\SearchResultWidget
is super-specialised, based on bespoke technologies, and only for use on
Special:Search, is not great.

None of these are used outside the MediaWiki core codebase and so this change
whilst it would otherwise be technically breaking is fine.

Change-Id: I16d21dd5e1ac3aa9893863df272d1bbb4788066f