ImageHistoryPseudoPager: Update doQuery() for IndexPager changes
This class duplicates a bunch of code from IndexPager, and that code was
changed in 6786aa5d8e2e87f00a3b1f81d9684ff4eac676ef in a way that broke
this code. Update it to account for the fact that mFirstShown,
mLastShown and mPastTheEndIndex are now arrays. Also update the
documentation in IndexPager to reflect that.

Bug: T244937
Change-Id: I51a50b6d3be1467f4ee399446d1d12cfed71a06c