SpecialWhatLinksHere: Fix queries (2020 edition)
The query being used when 'hidelinks' is set but 'hideredirs' is not was
not behaving well on pages with a very large number of pagelinks, as the
STRAIGHT_JOIN in the query was forcing it to scan all the pagelinks rows
to find the handful that were redirects.

The STRAIGHT_JOIN was added in I825eedac to avoid another stupid plan,
where MySQL decided to scan the whole of the `page` table then filter by
`redirect` and `pagelinks`. But the need for the `page` table in there
at all was removed when I4ccd7dfe removed $wgUseLinkNamespaceDBFields.
So we can get rid of the `page` join, which should let us remove the
STRAIGHT_JOIN without reopening T106682.

But we can also do one better: If 'hidelinks' is set we don't need to
check `pagelinks` at all, we can use `redirect` alone to get the
information we need directly.

Bug: T144235
Change-Id: Ida5f2fee548215cc190c401c2677c943f7bfe87b