resourceloader: Speed up structure/ResourcesTest
Avoid data providers for dynamic expansion over all JS, CSS
and images files known to MW and extensions (1000+), because
there's non-trivival overhead in PHPUnit for creating a separate
"test" wrapper and iteration for each.

Instead, make each group a single test with each of the checks
being an assertion within that.

Reduces test count of 'php phpunit.php structure/ResourcesTest.php'
(with MW core + Vector + Minerva installed) from 1039 tests to
37 tests, and reduces runtime from ~1.6min to ~40sec.

This is the same pattern I previously adapted in a5a672b1a5fd
for the biggest part test of this suite (dependency verification),
but I forgot to apply it to all the others.

Bug: T225730
Change-Id: Iae46b2d9ff8cd70aa6e64aaddf6ccf839591153b