mediawiki.user: Switch sessionId() back to session-cookie persistence
This change was tested in MediaWiki-Vagrant. The identifier is now
correctly shared by multiple tabs to the same website and clears
when the browser closes (and "restore session" is not used). The
previous behavior of persisting through multiple page views within
the same tab is retained. This change reverts a change from 2017.

Also, per JL's suggestion the comment about not naming cookies a
certain way (e.g. "[Ss]ession") was removed since it is Foundation-
specific due to varnish (T101857), and thus does not belong in Core.

Bug: T223931
Change-Id: I8448293b22acd4c4bea0791bfcc84d39ef4e1e5e