Add Parsoid as a library dependency
As Parsoid (including its integration with core) is still in active
development, we are using a '^0.12.0' version constraint for now.
This allows us to update Parsoid on both the production cluster and
the Parsoid cluster by updating the mediawiki-vendor repository
(composer.json/composer.lock/etc) without requiring further weekly
changes to core's `composer.json`.  We will convert this loose
constraint to a standard precise constraint when we branch core for
release.

This commit just adds parsoid as a library.  A follow-on commit
(Ie336f69692317d7a714809b42e035fb188e66fc0) then removes the auto-
loading of Parsoid from the Parsoid extension.

This patch doesn't (yet) create any specific uses of Parsoid from
the core codebase, but that is expected to follow rapidly once
we work through this initial deploy/release process hurdle.

Bug: T213494
Bug: T239660
Bug: T240055
Depends-On: I4c0367d6424513767751a2fc1dbe882250bef242
Change-Id: I8b3d3591b9a4f2a48ac83e1d242755c43b223aba