Remove wfGetMessageCacheStorage
There were no callers to this method, outside the $messageMemc
global, which is to be deprecated with T243175. Pending its
eventual removal, its usage has been inlined as the function
was trivial anyways.

Any previous callers were migrated with the convertion of the
MessageCache to a service in 752e7dd707a9.

Bug: T243176
Change-Id: I79846442512f023dcdf2f1f8c425156589c0421d