DevelopmentSettings: Enable 'fatal' and 'DeferredUpdates' error channels
These used to be in the 'exception' channel but were separated out
to better categorise them in Logstash, but this meant we regressed
in terms of CI enforcement because they weren't enabled (they only
went to the debug file), and thus Jenkins didn't fail the build
if they were non-empty.

To avoid having to keep this in sync with the Jenkins job config,
standardise on a single file (mw-error.log), which the job already
asserts as having to be empty/non-existent.

Change-Id: I432682c082f77d9a6ec2ea07539c2a3e41ea1340