mediawiki.searchSuggest: Inline use of serializeObject()
This random method exists in the jquery.getAttrs module because
the serializeArray() method for HTML forms, returns an array that
contains plain objects containing 'name' and 'value' properties,
which just so happens to be duck-type-compatible with the Node.attributes
interface on DOM nodes which lists AttributeNode objects that
also have 'key' and 'name' properties.

This is only ever used in searchSuggest.js, so simply inline it
there. This will be removed from getAttrs in the next commit.

Bug: T245894
Change-Id: I271c8f0fa9d8370d8fb33c9cae1e941fff9ee961