include blob: as a default script-src
blob: is basically a fancy version of eval(), commonly used with
web workers. We currently allow eval() ('unsafe-eval'), so we
might as well include blob. If we try and lock-down eval() at
some later point, we can re-evaluate this decision.

AFAIK, blob: urls are currently used with web workers in
NavigationTiming and CodeEditor extensions.

Bug: T245981
Change-Id: I3c44286e140ddfe2d3abd31428e0770ff5531e37