MediaWikiVersionFetcher: Change regex to capture any quoted string
Deciding on the shape of the contents of this string seems out of
scope for this class. It just needs to return what is in the string
and make sure to match anything that isn't the string value of
the MW_VERSION constant. That means we need to capture from quote
to quote.

Change-Id: Iee734f8cb425be49601ee4211bb9c01f04164a3f