TemplateParser: Make cache value include metadata
Update TemplateParser::compile to return the PHP code and metadata. The
metadata includes details about the compilation that can be used to
validate that PHP code fetched from a cache is fresh without having to
recompile the original template.

TemplateParser::getTemplate did have such an invalidation mechanism but:

- It stored the hash of the template file in the key

- It stored the hash used for checking the integrity of the cached PHP
  code in the first 64 bytes of the value

This tight coupling between the compilation result and the contents of
the cache key/value made changing either the compilation or cache
invalidation steps difficult.

After this change, arbitrary metadata can be added or removed from the
compilation result.

Changes:

- Move all filename generation and reading to TemplateParser::compile
  and align its signature with ::getTemplate

- Update TemplateParser::compile to generate the file hash with
  FileContentsHasher::getFileContentsHash, which is used by
  Resource Loader to generate file hashes, and return it as part of its
  result

- Update TemplateParser::getTemplate to store the integrity hash in the
  compilation result prior to caching it

- Bump the cache key major version as both the key and value have
  changed

Note well that TemplateParser isn't expected to be subclassed, and
::getTemplate and ::compile are protected members and are therefore out
of the scope of MediaWiki's deprecation policy.

Bug: T113095
Change-Id: Ifb5e122c6ae238fd300cd60f2b1ab33b7ece8e71