editpage: Set tab index on "Cancel" button that follows the other buttons
When page loads, cancel button doesn't get
focus when user hits the tab button continuously.
Hence tabbing is not intuitive to keyboard only users.

Adding a tabindex property to the cancel button to force the button to
belong in the same tab order as "Save Changes",  "Show Preview" and
"Show Changes" button .

Bug: T246257
Change-Id: I38e6c6e047ce93d7101e91d1b94c198f0ff7a19a