filerepo: Add and fix visibility keywords in ArchivedFile
The constructor should be a no-brainer, because it is actually used
in several places outside of this class.

The $exists property is used in the Video extension:
https://codesearch.wmflabs.org/search/?q=extends%20ArchivedFile
This extension is broken right now.

There is no obvious workaround, as far as I can see. I noticed that
almost all other properties are also protected. So why not make this
protected as well?

Change-Id: Ie80360e6362cd7fd4f3d9c9a60b6f64e8e9532f0