mw.Api: Add utility function for asserting the current user
Accidentally recording IP addresses in public change logs because
JS code used postWithToken is a common problem. (Normally when
due to session expiry or a similar reason the user is not logged
in anymore, actions fail with a CSRF error but postWithToken
fetches fresh tokens and avoids that.)

To make it easier to avoid (and mainly to make code doing that
easier to read), add a utility function for extending an API
request with assertions that make the request fail if the
client-side logic and the server disagree on the identity
of the user.

Change-Id: I98229796539e201f7a515b4d8ec68db8e30da53d