Allow adjusting CSP headers from BeforePageDisplay hook
Previously this hook ran after the header was sent, so it was
very confusing why CSP adjustments were being ignored by it. This
was especially the case, as it is a very natural point to adjust
the CSP policy, as commonly it is used to adjust what modules are
loaded.

Bug: T246614
Change-Id: I38f8f43fdc1d9bbd2f9af04305aad7a370b107b6