Send restrictive CSP header on things that do custom output
The idea of this patch, is to disable javascript for various weird
error outputs that don't use OutputPage. Thus if there is an XSS in
any of them, it is useless. Additionally, pages without a CSP policy
can be used to bypass network location restrictions, so this is a
step in the direction of locking that down.

Note: The CSP header only applies when using the page as a document,
it should have no effect if using the page as a subresource of
some other document.

Bug: T246639
Change-Id: I41f1f836eb8e52e72f0832bf74d8f75618947c8d