Add unit tests for SiteConfiguration extraction with nested strings
The $lang and $site placeholder variables don't just work on
strings assigned to wg-globals, they also work on strings within
arrays.

Add a test for this because it turns out this part is the slowest
part of the entire extraction process despite the fact that of the
thousands of variables, only a handful use these placeholders,
and even fewer use them within an array.

Bug: T169821
Change-Id: I1706a95764aa7d0ebfa627fb353fa496ccda5eee