parser: Fix incomplete PHPDocs, remove meaningless comments
As long as no parsing was started, and $this->startParse() was not
triggered at least once, the $this->mOptions property is null. This
does have a lot of consequences. I'm intentionally not fixing all of
them in this patch to not make it to big, only the most obvious ones.

This came up in I6714400 which was actually running into this issue.

This patch also removes a few comments that *literally* repeat what
the code already says. Such comments are not helpful, but clutter the
code, make it harder to read, and create error-prone redundancy.

Change-Id: Ic1ab18242b2f79d3da2a2e7c25e3af3e6907f7b4