Make Title::__construct private
The method was already marked as protected. This, together with the
Stable interface policy [1], means that it can be treated as unstable
and changed without prior notification.

In the future, we might want to make the Title class only instantiable
by TitleFactory (T247190). The first step is making the constructor
private, so that we can safely change the signature later.

[1] - https://www.mediawiki.org/wiki/Stable_interface_policy

Change-Id: I65830616ce584e9a553b93d99d58f8ecb02237b4